feat: api error handling
