+++ package.json
@@ -39,6 +39,7 @@
     "dotenv": "^16.0.3",
     "ethers": "^5.7.2",
     "express": "^4.18.2",
+    "express-async-errors": "^3.1.1",
     "multer": "^1.4.5-lts.1",
     "npm-run-all": "^4.1.5",
     "pluralistic": "github:gitcoinco/pluralistic.js#01e869460cd3e14637fea9fd55f3fe0540c73660",
