+++ src/expressAsyncErrors.d.ts
@@ -1,0 +1,2 @@
+// this module doesn't have any types
+declare module "express-async-errors";
